[Update]修正

+++ db/migrate/20220709145226_create_mangas.rb
@@ -5,7 +5,7 @@ class CreateMangas < ActiveRecord::Migration[6.1]
       t.string :title,null: false
       t.string :author,null: false
       t.string :rakuten,null: false
-      t.integer :isbn
+      t.decimal :isbn
       t.text :rakuten_image_url
 
       t.timestamps
